Renamed hello.go to main.go, add error handling, print full robots.txt
